backend/requirements.txt
fastapi[all]
uvicorn
pymongo
paho-mqtt
